Feature: Day Parting
Day Parting, which is described in
bat-native-ads/data/resources/catalog-scheme.json
is now added as a filter rule. Any existing fields with empty parting
will not be filtered.

The major note for server-side devs and reviewers is that the `dow`
field will store the C++ version of day of week as a string of integers,
where 0 is Sunday and 6 is Saturday. This is done to compress the catalog size
for multiple days using the same time period.